seperate signal func, set exit status after waitpid
