+++ src/main.c
@@ -1,23 +1,5 @@
 #include "../includes/minishell.h"
 
-void		sig_handler(int signo)
-{
-	char	buffer[MAXPATHLEN];
-	char	*current_path;
-
-	if (signo == SIGINT)
-	{
-		current_path = getcwd(buffer, MAXPATHLEN);
-		ft_putstr("\b\b  \b\b\n");
-		ft_putstr("catshell@");
-		ft_putstr_fd(buffer, 1);
-		ft_putstr(" ");
-		set_exit_status(1);
-	}
-	if (signo == SIGQUIT)
-		ft_putstr("\b\b  \b\b");
-}
-
 /*
 **	crtl + c (EOF)가 입력될 때 까지 get_next_line을 실행하여
 **  문자열을 입력받고, 입력 받은 문자열을 출력해주었다.
